requirements.txt
pytest~=8.3.3
selenium~=4.25.0
python-dotenv~=1.0.1
webdriver-manager~=4.0.2
requests~=2.32.3
colorama~=0.4.6
